+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping
 
 import woowacourse.shopping.di.AppModule
-import woowacourse.shopping.hashdi.DiApplication
+import woowacourse.shopping.hasydi.DiApplication
 
 class ShoppingApplication : DiApplication(listOf(AppModule()))
