+++ app/src/main/java/woowacourse/shopping/di/AppModule.kt
@@ -8,10 +8,10 @@ import woowacourse.shopping.data.ProductDefaultRepository
 import woowacourse.shopping.data.ShoppingDatabase
 import woowacourse.shopping.di.qualifier.InMemory
 import woowacourse.shopping.di.qualifier.RoomDB
-import woowacourse.shopping.hashdi.Module
-import woowacourse.shopping.hashdi.annotation.ApplicationContext
-import woowacourse.shopping.hashdi.annotation.Inject
-import woowacourse.shopping.hashdi.annotation.Singleton
+import woowacourse.shopping.hasydi.Module
+import woowacourse.shopping.hasydi.annotation.ApplicationContext
+import woowacourse.shopping.hasydi.annotation.Inject
+import woowacourse.shopping.hasydi.annotation.Singleton
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
