+++ app/src/main/java/woowacourse/shopping/di/CartActivityModule.kt
@@ -1,10 +1,10 @@
 package woowacourse.shopping.di
 
 import android.content.Context
-import woowacourse.shopping.hashdi.Module
-import woowacourse.shopping.hashdi.annotation.ActivityContext
-import woowacourse.shopping.hashdi.annotation.Inject
-import woowacourse.shopping.hashdi.annotation.Singleton
+import woowacourse.shopping.hasydi.Module
+import woowacourse.shopping.hasydi.annotation.ActivityContext
+import woowacourse.shopping.hasydi.annotation.Inject
+import woowacourse.shopping.hasydi.annotation.Singleton
 import woowacourse.shopping.ui.cart.DateFormatter
 
 class CartActivityModule : Module {
