+++ app/src/main/java/woowacourse/shopping/di/qualifier/InMemory.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.di.qualifier
 
-import woowacourse.shopping.hashdi.annotation.Qualifier
+import woowacourse.shopping.hasydi.annotation.Qualifier
 
 @Qualifier
 annotation class InMemory
