--- hasydi/src/main/java/woowacourse/shopping/hashdi/Module.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.hashdi
-
-interface Module
