--- hasydi/src/main/java/woowacourse/shopping/hashdi/annotation/ActivityContext.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.hashdi.annotation
-
-annotation class ActivityContext
