+++ hasydi/src/main/java/woowacourse/shopping/hasydi/DiApplication.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.hashdi
+package woowacourse.shopping.hasydi
 
 import android.app.Application
