+++ hasydi/src/main/java/woowacourse/shopping/hasydi/Module.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.hasydi
+
+interface Module
