+++ hasydi/src/main/java/woowacourse/shopping/hasydi/annotation/ActivityContext.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.hasydi.annotation
+
+annotation class ActivityContext
