+++ hasydi/src/main/java/woowacourse/shopping/hasydi/annotation/ApplicationContext.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.hasydi.annotation
+
+annotation class ApplicationContext
