+++ hasydi/src/main/java/woowacourse/shopping/hasydi/annotation/Qualifier.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.hashdi.annotation
+package woowacourse.shopping.hasydi.annotation
 
 @Target(AnnotationTarget.ANNOTATION_CLASS)
 annotation class Qualifier
